fix: Refactor CLI, fix example + path
