+++ packages/create-toolpad-app/src/package.ts
@@ -1,0 +1,27 @@
+import path from 'path';
+import * as fs from 'fs/promises';
+import type { PackageManager } from './types';
+
+export function getPackageManager(): PackageManager {
+  const userAgent = process.env.npm_config_user_agent;
+
+  if (userAgent) {
+    if (userAgent.startsWith('yarn')) {
+      return 'yarn';
+    }
+    if (userAgent.startsWith('pnpm')) {
+      return 'pnpm';
+    }
+    if (userAgent.startsWith('npm')) {
+      return 'npm';
+    }
+  }
+  return 'pnpm';
+}
+
+export async function findCtaPackageJson() {
+  const ctaPackageJsonPath = path.resolve(__dirname, '../package.json');
+  const content = await fs.readFile(ctaPackageJsonPath, 'utf8');
+  const packageJson = JSON.parse(content);
+  return packageJson;
+}
